% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTb.R
\name{combTb}
\alias{combTb}
\title{Cives the all opposite side of a vector}
\usage{
combTb(tb)
}
\arguments{
\item{tb}{vector}
}
\value{

}
\description{
Cives the all opposite side of a vector
}
\examples{

tb <- runif(2)/10
tbl <- combTb(tb)

X <- matrix(c(1,1),nrow = 1)


plot(X)
for(i in 1:nrow(tbl)){
lines(X + tbl[i,],type = "p",col = i)
}

}
