% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Chech if a vector is the same of another}
\usage{
compare(u, v, eps = 1e-09)
}
\arguments{
\item{eps}{}
}
\value{

}
\description{
Chech if a vector is the same of another
}
\examples{
compare(c(1,1),c(1,1))
compare(c(1,1),c(1,2))
}
