% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\title{Distance matrix}
\usage{
distMat(X, tore = FALSE, toreBound = 0)
}
\arguments{
\item{X}{Coordinates data.frame}
}
\value{
distance matrix
}
\description{
Return the distance matrix.
}
\examples{
N <- 36
X <- as.matrix(expand.grid(1:sqrt(N),1:sqrt(N)))
distMat(X,tore = TRUE,toreBound = 6)

}
