% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distRef.R
\name{distRef}
\alias{distRef}
\title{Distance from ref}
\usage{
distRef(X, ref, tb, tore = FALSE, toreBound = 0)
}
\arguments{
\item{X}{Coordinates data.frame}
}
\value{
distance matrix
}
\description{
Return the refth row of the matrix
}
\examples{
N <- 36
X <- as.matrix(expand.grid(1:sqrt(N),1:sqrt(N)))
distRef(X,ref = 4,tb = rep(0,2),tore = TRUE,toreBound = sqrt(N))

}
