% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodStrata.R
\name{goodStrata}
\alias{goodStrata}
\title{Find the strata that have the smallest distance at the end}
\usage{
goodStrata(u_Strata)
}
\arguments{
\item{X}{sptaial coordinates}

\item{pik_tmp}{temporary pik (unit k has a cutted pik)}

\item{toreBound}{}
}
\value{

}
\description{
Find the strata that have the smallest distance at the end
}
\examples{

############################# 1D

rm(list = ls())
set.seed(1)
eps <- 1e-13
N <- 10
n <- 3
X <- as.matrix(seq(1,N,1))
tb = 0.001
pik <- rep(0.3,N)
ref <- 6
pik[ref] <- runif(1,min = eps,max = pik[ref] - eps)
tore = TRUE
toreBound = 10
l_Strata <- allStrata(X,pik,ref,tb)
u_Strata <- unionStrata(l_Strata,ref)
s <- goodStrata(u_Strata)
plot(s,X)

################################## 2D

rm(list = ls())
set.seed(1)
eps <- 1e-13
N <- 36
n <- 8
pik <- rep(n/N,N)


tb <- runif(2)/100
X <- as.matrix(cbind(runif(N),runif(N)))
tore = FALSE
toreBound = FALSE
ref = 1
pik[ref] <- runif(1,min = eps,max = pik[ref] - eps)

l_Strata <- allStrata(X,pik,ref,tb)
u_Strata <- unionStrata(l_Strata)
s <- goodStrata(u_Strata)
plot(s,X)
}
