% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderRef.R
\name{orderRef}
\alias{orderRef}
\title{Reorder the strata from ref perspective}
\usage{
orderRef(strata, d_ref, ref)
}
\arguments{
\item{d_ref}{}
}
\value{

}
\description{
Reorder the strata from ref perspective
}
\examples{
strata <- c(4, 5, 6)
d_ref <- c( 3.996001, 0.998001, 0.000000)
ref = 6
orderRef(strata,d_ref,ref)

strata <- c(6, 7, 8, 1, 2)
d_ref <- c(0.000000,  0.998001,  3.996001,  8.994001, 15.992001)
ref = 6
orderRef(strata,d_ref,ref)

}
