% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{Title}
\usage{
simu(SIM, Xcoord, pik, tb, tore, toreBound)
}
\arguments{
\item{toreBound}{}
}
\value{

}
\description{
Title
}
\examples{

############# 2D runif
rm(list = ls())
N <- 36
n <- 6
Xcoord <- as.matrix(cbind(runif(N, min = 1,max = 10),runif(N,min = 1,max = 10)))
pik <- rep(n/N,N)
tb <- rep(0,2)
tore = FALSE
toreBound = -1
SIM <- 200

res <- simu(SIM,Xcoord,pik,tb,tore,toreBound)

library(ggplot2)
dat <- data.frame(x = Xcoord[,1],y = Xcoord[,2],size = res$pik_systematic,pik = pik)

ggplot()+
  geom_point(data = dat,aes(x = x,y = y,size = pik),pch = 1,alpha = 1,colour = "black")+
  geom_point(data = dat,aes(x = x,y = y,size = size,colour = size))+
  theme_bw()


############# 2D grid
rm(list = ls())
N <- 36
n <- 6
Xcoord <- as.matrix(expand.grid(seq(1,sqrt(N),1),seq(1,sqrt(N),1)))
pik <- rep(n/N,N)
tore = TRUE
toreBound = 6
tb <- runif(2)/100
SIM = 100
s <- systematic(Xcoord,pik,tb,tore = tore,toreBound = toreBound,comment = TRUE)
plot(Xcoord)
points(Xcoord[s,],pch = 16)

res <- simu(SIM,Xcoord,pik,tb,tore,toreBound)

library(ggplot2)
dat <- data.frame(x = Xcoord[,1],y = Xcoord[,2],size = res$pik_systematic,pik = pik)

ggplot()+
  geom_point(data = dat,aes(x = x,y = y,size = pik),pch = 1,alpha = 1,colour = "black")+
  geom_point(data = dat,aes(x = x,y = y,size = size,colour = size))+
  theme_bw()

############# 1D
rm(list = ls())
N <- 30
n <- 5
Xcoord <- as.matrix(seq(1,N,1))
pik <- inclusionprobabilities(runif(N),n = 5)
tb = 0.001
tore <- TRUE
toreBound <- N
SIM = 1000
simu(SIM,Xcoord,pik,tb,tore,toreBound)

}
