% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataCompleteLink.R
\name{strataCompleteLink}
\alias{strataCompleteLink}
\title{Title}
\usage{
strataCompleteLink(
  X,
  pik,
  ref,
  k,
  tb,
  pikInit,
  bound = 1,
  tore = FALSE,
  toreBound = -1
)
}
\arguments{
\item{cum}{}
}
\value{

}
\description{
Title
}
\examples{

rm(list = ls())
set.seed(6)
eps <- 1e-13
N <- 36
n <-  6
pik <- rep(n/N,N)
pik <- inclusionprobabilities(runif(N),n)

tb <- runif(2,min = -0.001,max = 0.001)
tb
X <- as.matrix(expand.grid(seq(1,sqrt(N),1),seq(1,sqrt(N),1)))
tore = TRUE
toreBound = sqrt(N)
comment = TRUE
bound = 1

ref <- 16
s <- ref
k = 9
pik_cutted <-  runif(1,min = eps,max = pik[s[1]] - eps)
pikstar <- pik
pikstar[s] <- pik_cutted
pikInit <- pik
pik <- pikstar

strata <- strataCompleteLink(X,pik,ref,k,tb,pikInit)
strata <- strataCompleteLink(X,pik,ref,k,tb,pikInit)










}
