% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{systematic}
\alias{systematic}
\title{Title}
\usage{
systematic(
  X,
  pik,
  tb,
  tore = FALSE,
  toreBound = -1,
  comment = FALSE,
  pik_cutted,
  s
)
}
\arguments{
\item{X}{sptaial coordinates}

\item{toreBound}{}

\item{pik_tmp}{temporary pik (unit k has a cutted pik)}
}
\value{

}
\description{
La fonction va boucler sur les unités les plus proche pour trouver la strate tel que l'unité k est au bord
(coupé de la bonne manière) et un autre unité qui laisse le plus en dehors.
}
\examples{



############################# 1D

rm(list = ls())
eps <- 1e-13
N <- 36
n <- 8
X <- as.matrix(seq(1,N,1))
tb = -0.001
pik <- rep(n/N,N)
shift = TRUE
tore = TRUE
toreBound = N
comment = TRUE


s <- systematic(X,pik,tb,tore,toreBound,comment = TRUE)
plot(X,rep(0,N))
points(X[s],rep(0,n),pch = 16)

############################# 2D GRID UNEQUAL


rm(list = ls())
#set.seed(2)
# set.seed(8)
# set.seed(3)
 #set.seed(5) # n = 12
set.seed(7)
eps <- 1e-13
N <- 144
n <-  24
pik <- rep(n/N,N)
#pik <- sampling::inclusionprobabilities(runif(N),n)

tb <- runif(2)/100

X <- as.matrix(expand.grid(seq(1,sqrt(N),1),seq(1,sqrt(N),1)))
tore = TRUE
toreBound = sqrt(N)
comment = TRUE

s <- systematic(X,pik,tb,tore = tore,toreBound = toreBound,comment = comment)

############################# 2D GRID


rm(list = ls())
#set.seed(2)
# set.seed(8)
# set.seed(3)
 #set.seed(5) # n = 12
set.seed(7)
eps <- 1e-13
N <- 144
n <-  24
pik <- rep(n/N,N)


tb <- runif(2)/100

X <- as.matrix(expand.grid(seq(1,sqrt(N),1),seq(1,sqrt(N),1)))
tore = TRUE
toreBound = sqrt(N)
comment = TRUE

s <- systematic(X,pik,tb,tore = tore,toreBound = toreBound,comment = comment)

plot(X)
points(X[s,],pch = 16)


############################# 2D


rm(list = ls())
#set.seed(3)
eps <- 1e-13
N <- 100
n <- 10
pik <- rep(n/N,N)


tb <- rep(0,2)
X <- as.matrix(cbind(runif(N, min = 1,max = 3),runif(N,min = 1,max = 3)))
tore = FALSE
toreBound = -1
comment = TRUE
s <- systematic(X,pik,tb,tore = tore,toreBound = toreBound,comment = TRUE)
plot(X)
points(X[s,],pch = 16)

s <- lpm1(pik,X)
plot(X)
points(X[s,],pch = 16)

}
