% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionStrata.R
\name{unionStrata}
\alias{unionStrata}
\title{Create the union of the strata}
\usage{
unionStrata(l_Strata, ref)
}
\arguments{
\item{X}{sptaial coordinates}

\item{pik_tmp}{temporary pik (unit k has a cutted pik)}

\item{toreBound}{}
}
\value{

}
\description{
Create the union of the strata
}
\examples{


############################# 1D

rm(list = ls())
set.seed(1)
eps <- 1e-13
N <- 10
n <- 3
X <- as.matrix(seq(1,N,1))
tb = -0.001
pik <- rep(0.3,N)
pikInit <- pik
ref <- 6
pik[ref] <- runif(1,min = eps,max = pik[ref] - eps)
shift = TRUE
tore = TRUE
toreBound = 10
l_Strata <- allStrata(X,pik,ref,tb,pikInit)
l_Strata
u_Strata <- unionStrata(l_Strata,ref)
u_strata

################################# 2D
rm(list = ls())
set.seed(1)
N <- 36
n <- 8
pik <- rep(n/N,N)
pikInit <- pik


tb <- runif(2)/100
X <- as.matrix(cbind(runif(N),runif(N)))
tore = FALSE
toreBound = FALSE
ref = 1

l_Strata <- allStrata(X,pik,ref,tb,pikInit)
plot(l_Strata,X)
u_Strata <- unionStrata(l_Strata)
plot(u_Strata,X)

################## GRID 2D

rm(list = ls())
set.seed(1)
eps <- 1e-13
N <- 81
n <- 9
pik <- rep(n/N,N)


tb <- runif(2)/100

X <- as.matrix(expand.grid(seq(1,sqrt(N),1),seq(1,sqrt(N),1)))
tore = TRUE
toreBound = sqrt(N)

ref = 25

l_Strata <- allStrata(X,pik,ref,tb,tore = tore,toreBound = toreBound,pikInit = pik)
plot(l_Strata,X)
u_Strata <- unionStrata(l_Strata)
plot(u_Strata,X)
}
