% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allStrata.R
\name{allStrata}
\alias{allStrata}
\title{Find all strata that cover in any way the re unit}
\usage{
allStrata(
  X,
  pik,
  ref,
  tb,
  pikInit,
  bound = 1,
  multi = TRUE,
  tore = FALSE,
  toreBound = -1,
  comment = FALSE
)
}
\arguments{
\item{X}{sptaial coordinates}

\item{pik}{inclusion probabilies}

\item{ref}{the referenced unit}

\item{tb}{a shift}

\item{bound}{the bound for the sum of inclusion probabilites}

\item{tore}{a logcial that see the spatial coordinates on tore}

\item{toreBound}{the bound of the tore}

\item{comment}{}
}
\value{

}
\description{
Find all strata that cover in any way the re unit
}
