% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\title{Distance matrix}
\usage{
distMat(X, ref, tb, tore = FALSE, toreBound = 0)
}
\arguments{
\item{X}{Coordinates data.frame}
}
\value{
distance matrix
}
\description{
Return the distance matrix.
}
\examples{
N <- 36
X <- as.matrix(expand.grid(1:sqrt(N),1:sqrt(N)))
N <- nrow(X)
tb = runif(2)/100
#f <- function(k,Xcoord,tore,toreBound){distUnitk(as.matrix(Xcoord),k =k,tore, toreBound)}
#system.time(do.call(cbind,lapply(X = 1:N,FUN = f,Xcoord = X,tore = TRUE,toreBound = sqrt(N))))
system.time(D <- distMat(X,ref = 1,tb,tore = TRUE,toreBound = sqrt(N)))

}
